RSCAD IEEE 13 node
This is the model of the IEEE 13 node bus system for RSCAD V4 platform. The files are all that are required for model to be executed.

DISCLAIMER: The model is provided by RSCAD and is being used only for demonstration purposes. INL does not claim any part of the model to be its own property.